Aggiungi i serializer per Star e Planet con supporto per la distanza e aggiorna le view corrispondenti
