+++ eddai_EliteDangerousApiInterface/ed_body/api/venws/starViewSet.py
@@ -1,8 +1,10 @@
 from core.api.viewsets import OwnerAndDateModelViewSet
+from ed_core.api.mixins import DistanceModelMixin
+
 from rest_framework.filters import SearchFilter
 from django_filters.rest_framework import DjangoFilterBackend
 
-from ..serializers import StarSerializer
+from ..serializers import StarSerializer, StarDistanceSerializer
 from ..filterSet import StarFilterSet
 
 from ed_body.models import Star
@@ -19,6 +21,8 @@ class StarViewSet(OwnerAndDateModelViewSet):
     
     queryset = Star.objects.all()
     serializer_class = StarSerializer
+    distance_serializer_class = StarDistanceSerializer
+    filter_param_distance = 'order_by_system'
     filterset_class = StarFilterSet
     filter_backends = [SearchFilter, DjangoFilterBackend]
     search_fields = ['name']
