+++ eddai_EliteDangerousApiInterface/ed_core/api/mixins/__init__.py
@@ -1,0 +1,1 @@
+from .distanceModelMixin import DistanceModelMixin
